chore(refactor): improve option `exclude`
