+++ tools/compatibility/deno.ts
@@ -1,5 +1,5 @@
 import { promises as fs } from 'node:fs';
-import { getFiles } from '../../src/helpers/get-files.ts';
+import { getFiles } from '../../src/modules/get-files.ts';
 
 const ensureDenoCompatibility = async (path: string) => {
   const files = getFiles(path, [], {
